+++ dsaa/c/src/double_list.c
@@ -5,7 +5,7 @@
 void DoubleListInit(DoubleList *p_list, size_t unit_size, size_t hint_pool_size)
 {
 #if ENABLE_DSAA_OPTIMIZATION
-	MemoryPoolInit(&p_list->pool, hint_pool_size, unit_size + sizeof(DoubleListNode));
+	MemoryPoolInit(&p_list->pool, (unsigned int)hint_pool_size, (unsigned int)(unit_size + sizeof(DoubleListNode)));
 	p_list->head = (DoubleListNode*)MemoryPoolAlloc(&p_list->pool);
 	p_list->tail = (DoubleListNode*)MemoryPoolAlloc(&p_list->pool);
 #else
